+++ src/main/kotlin/no/nav/amt/distribusjon/hendelse/HendelseRepository.kt
@@ -4,12 +4,12 @@ import com.fasterxml.jackson.module.kotlin.readValue
 import kotliquery.Row
 import kotliquery.queryOf
 import no.nav.amt.distribusjon.application.plugins.objectMapper
+import no.nav.amt.distribusjon.db.toPGObject
 import no.nav.amt.distribusjon.distribusjonskanal.Distribusjonskanal
 import no.nav.amt.distribusjon.hendelse.model.Hendelse
 import no.nav.amt.distribusjon.journalforing.model.HendelseMedJournalforingstatus
 import no.nav.amt.distribusjon.journalforing.model.Journalforingstatus
 import no.nav.amt.lib.utils.database.Database
-import no.nav.amt.lib.utils.database.toPGObject
 import java.time.LocalDateTime
 import java.util.UUID
