+++ src/main/kotlin/no/nav/amt/distribusjon/hendelse/model/Tiltak.kt
@@ -1,11 +1,10 @@
 package no.nav.amt.distribusjon.hendelse.model
 
-import no.nav.amt.distribusjon.amtdeltaker.Tiltakstype
-
 data class Tiltak(
     val navn: String,
     val type: ArenaTiltakTypeKode,
     val ledetekst: String,
+    val tiltakskode: Tiltakskode?,
 )
 
 enum class ArenaTiltakTypeKode {
@@ -20,14 +19,26 @@ enum class ArenaTiltakTypeKode {
     GRUFAGYRKE,
 }
 
+enum class Tiltakskode {
+    ARBEIDSFORBEREDENDE_TRENING,
+    ARBEIDSRETTET_REHABILITERING,
+    AVKLARING,
+    DIGITALT_OPPFOLGINGSTILTAK,
+    GRUPPE_ARBEIDSMARKEDSOPPLAERING,
+    GRUPPE_FAG_OG_YRKESOPPLAERING,
+    JOBBKLUBB,
+    OPPFOLGING,
+    VARIG_TILRETTELAGT_ARBEID_SKJERMET,
+}
+
 fun ArenaTiltakTypeKode.toTiltakskode() = when (this) {
-    ArenaTiltakTypeKode.INDOPPFAG -> Tiltakstype.Tiltakskode.OPPFOLGING
-    ArenaTiltakTypeKode.ARBFORB -> Tiltakstype.Tiltakskode.ARBEIDSFORBEREDENDE_TRENING
-    ArenaTiltakTypeKode.AVKLARAG -> Tiltakstype.Tiltakskode.AVKLARING
-    ArenaTiltakTypeKode.ARBRRHDAG -> Tiltakstype.Tiltakskode.ARBEIDSRETTET_REHABILITERING
-    ArenaTiltakTypeKode.DIGIOPPARB -> Tiltakstype.Tiltakskode.DIGITALT_OPPFOLGINGSTILTAK
-    ArenaTiltakTypeKode.VASV -> Tiltakstype.Tiltakskode.VARIG_TILRETTELAGT_ARBEID_SKJERMET
-    ArenaTiltakTypeKode.JOBBK -> Tiltakstype.Tiltakskode.JOBBKLUBB
-    ArenaTiltakTypeKode.GRUPPEAMO -> Tiltakstype.Tiltakskode.GRUPPE_ARBEIDSMARKEDSOPPLAERING
-    ArenaTiltakTypeKode.GRUFAGYRKE -> Tiltakstype.Tiltakskode.GRUPPE_FAG_OG_YRKESOPPLAERING
+    ArenaTiltakTypeKode.INDOPPFAG -> Tiltakskode.OPPFOLGING
+    ArenaTiltakTypeKode.ARBFORB -> Tiltakskode.ARBEIDSFORBEREDENDE_TRENING
+    ArenaTiltakTypeKode.AVKLARAG -> Tiltakskode.AVKLARING
+    ArenaTiltakTypeKode.ARBRRHDAG -> Tiltakskode.ARBEIDSRETTET_REHABILITERING
+    ArenaTiltakTypeKode.DIGIOPPARB -> Tiltakskode.DIGITALT_OPPFOLGINGSTILTAK
+    ArenaTiltakTypeKode.VASV -> Tiltakskode.VARIG_TILRETTELAGT_ARBEID_SKJERMET
+    ArenaTiltakTypeKode.JOBBK -> Tiltakskode.JOBBKLUBB
+    ArenaTiltakTypeKode.GRUPPEAMO -> Tiltakskode.GRUPPE_ARBEIDSMARKEDSOPPLAERING
+    ArenaTiltakTypeKode.GRUFAGYRKE -> Tiltakskode.GRUPPE_FAG_OG_YRKESOPPLAERING
 }
