+++ src/main/kotlin/no/nav/amt/distribusjon/journalforing/pdf/HovedvedtakPdfDto.kt
@@ -1,6 +1,6 @@
 package no.nav.amt.distribusjon.journalforing.pdf
 
-import no.nav.amt.distribusjon.amtdeltaker.Tiltakstype
+import no.nav.amt.distribusjon.hendelse.model.Tiltakskode
 import java.time.LocalDate
 
 data class HovedvedtakPdfDto(
@@ -24,7 +24,7 @@ data class HovedvedtakPdfDto(
 
     data class DeltakerlisteDto(
         val navn: String,
-        val tiltakskode: Tiltakstype.Tiltakskode,
+        val tiltakskode: Tiltakskode,
         val ledetekst: String,
         val arrangor: ArrangorDto,
         val forskriftskapittel: Int,
